pyproject.toml
[tool.poetry]
name = "digital_currency_report"
version = "0.1.0"
description = ""
authors = ["Martin <1403951401@qq.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
uvloop = "^0.14.0"
python-rapidjson = "^1.0"
alembic = "^1.5.0"
sqlalchemy = "1.4.0b1"
asyncpg = "^0.21.0"
loguru = "^0.5.3"
requests-async = "^0.6.2"
aioredis = "^1.3.1"
gunicorn = "^20.0.4"
httptools = "^0.1.1"
python-multipart = "^0.0.5"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
